+++ changelog.md
@@ -5,6 +5,7 @@
 - Serial: fix b64/not b64 error
 - Serial: perf improvement
 - Base64: perf improvement
+- new: Proxy selector
 
 ## Version 4.5.1